% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_design.R
\name{get_design}
\alias{get_design}
\title{get_design

`get_design()` returns design information for optimal algorithm for 
generalized hierarchical group testing screening using dynamic programming.}
\usage{
get_design(p = NULL, df = NULL, id_var = NULL, p_var = NULL,
  Se = 1, Sp = 1, monte_carlo = TRUE, M = 10000)
}
\arguments{
\item{p}{Vector of prevalences for population. Optional, but either p or df 
must be given.}

\item{df}{data frame containing id variables and prevalences for all members
of population.}

\item{id_var}{Name of variable in df containing id information. Used to
name id variable in result.}

\item{p_var}{Name of variable in df containing prevalence information. Used
to name prevalence variable in result.}

\item{Se}{Assay sensitivity; assumed constant across group sizes.}

\item{Sp}{Assay specificity; assumed constant across group sizes.}

\item{monte_carlo}{A logical value indicating whether Monte Carlo methods
should be used to estimate overall sensitivity and overall
specificity..}

\item{M}{Number of Monte Carlo iterations.}
}
\value{
Returns a list with several design elements. 
  1. A data frame containing, for each member of the population, an id
  value (this is the same as the orginal population id if provided),
  prevalence value (name is p_var, defaults to p), the individuals rank
  in list of ordered prevalences (order), initial group assignments (group).
  Note, the group labels are such that all individals with the same number
  should be tested togethor in the first stage of the algorithm.

 2. The expected number of tests (ET) for completing the screening algorithm.
 
 3-4. The overall sensitivity (Se) and specificity (Sp). These are provided
      only if monte_carlo = TRUE.

 5. A function (get_next_group) which indicated how a group should be divided
    in subsequent stages following a positive group test. This function takes
    two arguments, index_1 and index_2, representing population rank of the
    individuals with the smallest and largest prevalences, respectively, in
    the positive group. These ranks can be obtained directly from the order
    column of the data frame returned with the name initial_design (the first
    list component).
}
\description{
get_design

`get_design()` returns design information for optimal algorithm for 
generalized hierarchical group testing screening using dynamic programming.
}
