% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{return_hdp}
\alias{return_hdp}
\title{return_hdp
constructs optimal hierarchical group testing design for ordered population
and returns expected number of tests}
\usage{
return_hdp(q, Se, Sp)
}
\arguments{
\item{q}{(Rcpp::NumericVector): ordered vector of individual prevalences.}

\item{Se}{(double): assay sensitivity.}

\item{Sp}{(double): assay specificity.}
}
\description{
return_hdp
constructs optimal hierarchical group testing design for ordered population
and returns expected number of tests
}
