% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{return_hdp_mc}
\alias{return_hdp_mc}
\title{return_hdp_mc
constructs optimal hierarchical group testing design for ordered population
and uses Monte Carlo to estimate the expected numnber of tests, overall 
sensitivity, and overall specificity.}
\usage{
return_hdp_mc(q, Se, Sp, M)
}
\arguments{
\item{q}{(Rcpp::NumericVector): ordered vector of individual prevalences.}

\item{Se}{(double): assay sensitivity.}

\item{Sp}{(double): assay specificity.}

\item{M}{(int): number of Monte Carlo iterations}
}
\description{
return_hdp_mc
constructs optimal hierarchical group testing design for ordered population
and uses Monte Carlo to estimate the expected numnber of tests, overall 
sensitivity, and overall specificity.
}
