% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_screening.R
\name{simulate_screening}
\alias{simulate_screening}
\title{simulate_screening

Simulates screening process for given dataset.}
\usage{
simulate_screening(x, p, Se, Sp)
}
\arguments{
\item{x}{Vector of individual statuses (binary).}

\item{p}{Vector of individual prevalence estimates.}

\item{Se}{Assay sensitivity.}

\item{Sp}{Assay specificity.}
}
\value{
List given screening results, number of tests required, overall
  sample sensitivity, and overall sample specificity.
}
\description{
simulate_screening

Simulates screening process for given dataset.
}
